OZ-562: Refactored component-specific user actions
