+++ e2e/tests/openmrs-senaite-flows.spec.ts
@@ -1,72 +1,67 @@
 import { test, expect } from '@playwright/test';
-import { HomePage } from '../utils/functions/testBase';
-import { patientName } from '../utils/functions/testBase';
+import { OpenMRS, patientName } from '../utils/functions/openmrs';
+import { SENAITE } from '../utils/functions/senaite';
 import { O3_URL } from '../utils/configs/globalSetup';
 
-let homePage: HomePage;
+let openmrs: OpenMRS;
+let senaite: SENAITE;
 
 test.beforeEach(async ({ page }) => {
-  homePage = new HomePage(page);
+  openmrs = new OpenMRS(page);
+  senaite = new SENAITE(page);
+
+  await openmrs.initiateLogin();
+  await expect(page).toHaveURL(/.*home/);
+  await openmrs.createPatient();
+  await openmrs.startPatientVisit();
 });
 
 test('Ordering a lab test for an OpenMRS patient creates the corresponding SENAITE client with an analysis request.', async ({ page }) => {
   // replay
-  await homePage.initiateLogin();
-  await expect(page).toHaveURL(/.*home/);
-  await homePage.createPatient();
-  await homePage.startPatientVisit();
-  await homePage.goToLabOrderForm();
+  await openmrs.goToLabOrderForm();
   await page.getByRole('button', { name: 'Add', exact: true }).click();
   await page.locator('#tab select').selectOption('857AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
-  await homePage.saveLabOrder();
+  await openmrs.saveLabOrder();
 
   // verify
-  await homePage.goToSENAITE();
+  await senaite.goToSENAITE();
   await expect(page).toHaveURL(/.*senaite/);
-  await homePage.searchClientInSENAITE();
+  await senaite.searchClientInSENAITE();
   const client = await page.locator('table tbody tr:nth-child(1) td.contentcell.title div span a');
   await expect(client).toContainText(`${patientName.firstName + ' ' + patientName.givenName}`);
 });
 
 test('Editing the details of an OpenMRS patient with a synced lab order edits the corresponding SENAITE client details.', async ({ page }) => {
   // replay
-  await homePage.initiateLogin();
-  await expect(page).toHaveURL(/.*home/);
-  await homePage.createPatient();
-  await homePage.startPatientVisit();
-  await homePage.goToLabOrderForm();
+  await openmrs.goToLabOrderForm();
   await page.getByRole('button', { name: 'Add', exact: true }).click();
   await page.locator('#tab select').selectOption('857AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
-  await homePage.saveLabOrder();
-  await homePage.goToSENAITE();
+  await openmrs.saveLabOrder();
+  await senaite.goToSENAITE();
   await expect(page).toHaveURL(/.*senaite/);
-  await homePage.searchClientInSENAITE();
+  await senaite.searchClientInSENAITE();
   const client = await page.locator('table tbody tr:nth-child(1) td.contentcell.title div span a');
   await expect(client).toContainText(`${patientName.firstName + ' ' + patientName.givenName}`);
   await page.goto(`${O3_URL}`);
-  await homePage.searchPatient(`${patientName.firstName + ' ' + patientName.givenName}`);
-  await homePage.updatePatientDetails();
+  await openmrs.searchPatient(`${patientName.firstName + ' ' + patientName.givenName}`);
+  await openmrs.updatePatientDetails();
 
   // verify
-  await homePage.goToSENAITE();
-  await homePage.searchClientInSENAITE();
+  await senaite.goToSENAITE();
+  await senaite.searchClientInSENAITE();
 
   await expect(client).toContainText(`${patientName.updatedFirstName}` + ' ' + `${patientName.givenName }`);
 });
 
 test('Editing a synced OpenMRS lab order edits the corresponding SENAITE analysis request.', async ({ page }) => {
   // replay
-  await homePage.initiateLogin();
-  await expect(page).toHaveURL(/.*home/);
-  await homePage.createPatient();
-  await homePage.startPatientVisit();
-  await homePage.goToLabOrderForm();
+  await openmrs.goToLabOrderForm();
   await page.getByRole('button', { name: 'Add', exact: true }).click();
   await page.locator('#tab select').selectOption('857AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
-  await homePage.saveLabOrder();
-  await homePage.goToSENAITE();
+  await openmrs.saveLabOrder();
+  await senaite.goToSENAITE();
   await expect(page).toHaveURL(/.*senaite/);
-  await homePage.searchClientInSENAITE();
+  await senaite.searchClientInSENAITE();
   const client = await page.locator('table tbody tr:nth-child(1) td.contentcell.title div');
   await expect(client).toContainText(`${patientName.firstName + ' ' + patientName.givenName}`);
   await page.locator('table tbody tr:nth-child(1) td.contentcell.title div').click();
@@ -74,12 +69,12 @@ test('Editing a synced OpenMRS lab order edits the corresponding SENAITE analysi
   const analysisRequest = await page.locator('#sampleheader-standard-fields tr:nth-child(1) td:nth-child(6)');
   await expect(analysisRequest).toHaveText('Blood urea nitrogen Template');
   await page.goto(`${O3_URL}`);
-  await homePage.searchPatient(`${patientName.firstName + ' ' + patientName.givenName}`);
-  await homePage.updateLabOrder();
+  await openmrs.searchPatient(`${patientName.firstName + ' ' + patientName.givenName}`);
+  await openmrs.updateLabOrder();
 
   // verify
-  await homePage.goToSENAITE();
-  await homePage.searchClientInSENAITE();
+  await senaite.goToSENAITE();
+  await senaite.searchClientInSENAITE();
 
   await expect(client).toContainText(`${patientName.firstName + ' ' + patientName.givenName}`);
   await page.locator('table tbody tr:nth-child(1) td.contentcell.title div').click();
@@ -89,17 +84,13 @@ test('Editing a synced OpenMRS lab order edits the corresponding SENAITE analysi
 
 test('Voiding a synced OpenMRS lab order cancels the corresponding SENAITE analysis request.', async ({ page }) => {
   // replay
-  await homePage.initiateLogin();
-  await expect(page).toHaveURL(/.*home/);
-  await homePage.createPatient();
-  await homePage.startPatientVisit();
-  await homePage.goToLabOrderForm();
+  await openmrs.goToLabOrderForm();
   await page.getByRole('button', { name: 'Add', exact: true }).click();
   await page.locator('#tab select').selectOption('857AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
-  await homePage.saveLabOrder();
-  await homePage.goToSENAITE();
+  await openmrs.saveLabOrder();
+  await senaite.goToSENAITE();
   await expect(page).toHaveURL(/.*senaite/);
-  await homePage.searchClientInSENAITE();
+  await senaite.searchClientInSENAITE();
   const client = await page.locator('table tbody tr:nth-child(1) td.contentcell.title div');
   await expect(client).toContainText(`${patientName.firstName + ' ' + patientName.givenName}`);
   await page.locator('table tbody tr:nth-child(1) td.contentcell.title div').click();
@@ -107,37 +98,33 @@ test('Voiding a synced OpenMRS lab order cancels the corresponding SENAITE analy
   const analysisRequest = await page.locator('#sampleheader-standard-fields tr:nth-child(1) td:nth-child(6)');
   await expect(analysisRequest).toHaveText('Blood urea nitrogen Template');
   await page.goto(`${O3_URL}`);
-  await homePage.searchPatient(`${patientName.firstName + ' ' + patientName.givenName}`);
-  await homePage.voidEncounter();
+  await openmrs.searchPatient(`${patientName.firstName + ' ' + patientName.givenName}`);
+  await openmrs.voidEncounter();
 
   // verify
-  await homePage.goToSENAITE();
-  await homePage.searchClientInSENAITE();
+  await senaite.goToSENAITE();
+  await senaite.searchClientInSENAITE();
   await expect(client).not.toHaveText(`${patientName.firstName + ' ' + patientName.givenName}`);
 });
 
 test('Published coded lab results from SENAITE are viewable in the OpenMRS lab results viewer.', async ({ page }) => {
   // replay
-  await homePage.initiateLogin();
-  await expect(page).toHaveURL(/.*home/);
-  await homePage.createPatient();
-  await homePage.startPatientVisit();
-  await homePage.goToLabOrderForm();
+  await openmrs.goToLabOrderForm();
   await page.getByRole('button', { name: 'Add', exact: true }).click();
   await page.locator('#tab select').selectOption('1325AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
-  await homePage.saveLabOrder();
-  await homePage.goToSENAITE();
+  await openmrs.saveLabOrder();
+  await senaite.goToSENAITE();
   await expect(page).toHaveURL(/.*senaite/);
-  await homePage.searchClientInSENAITE();
-  await homePage.createPartition();
+  await senaite.searchClientInSENAITE();
+  await senaite.createPartition();
   await page.getByRole('combobox', { name: 'Result' }).selectOption('664AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
-  await homePage.publishLabResults();
+  await senaite.publishLabResults();
   const reviewState = await page.locator('table tbody tr.contentrow.state-published.parent td.contentcell.State span span').textContent();
   await expect(reviewState?.includes('Published')).toBeTruthy();
 
   // verify
   await page.goto(`${O3_URL}`);
-  await homePage.viewTestResults();
+  await openmrs.viewTestResults();
   const testName = await page.locator('div:nth-child(2) >div> div.cds--data-table-container td:nth-child(1)').first();
   await expect(testName).toContainText('Hepatitis C test - qualitative');
 
@@ -147,26 +134,22 @@ test('Published coded lab results from SENAITE are viewable in the OpenMRS lab r
 
 test('Published numeric lab results from SENAITE are viewable in the OpenMRS lab results viewer.', async ({ page }) => {
   // replay
-  await homePage.initiateLogin();
-  await expect(page).toHaveURL(/.*home/);
-  await homePage.createPatient();
-  await homePage.startPatientVisit();
-  await homePage.goToLabOrderForm();
+  await openmrs.goToLabOrderForm();
   await page.getByRole('button', { name: 'Add', exact: true }).click();
   await page.locator('#tab select').selectOption('655AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
-  await homePage.saveLabOrder();
-  await homePage.goToSENAITE();
+  await openmrs.saveLabOrder();
+  await senaite.goToSENAITE();
   await expect(page).toHaveURL(/.*senaite/);
-  await homePage.searchClientInSENAITE();
-  await homePage.createPartition();
+  await senaite.searchClientInSENAITE();
+  await senaite.createPartition();
   await page.locator('tr:nth-child(1) td.contentcell.Result div span input').fill('64');
-  await homePage.publishLabResults();
+  await senaite.publishLabResults();
   const reviewState = await page.locator('table tbody tr.contentrow.state-published.parent td.contentcell.State span span').textContent();
   await expect(reviewState?.includes('Published')).toBeTruthy();
 
   // verify
   await page.goto(`${O3_URL}`);
-  await homePage.viewTestResults();
+  await openmrs.viewTestResults();
   const testName = await page.locator('div:nth-child(2) >div> div.cds--data-table-container td:nth-child(1)').first();
   await expect(testName).toContainText('Total bilirubin');
 
@@ -176,26 +159,22 @@ test('Published numeric lab results from SENAITE are viewable in the OpenMRS lab
 
 test('Published free text lab results from SENAITE are viewable in the OpenMRS lab results viewer.', async ({ page }) => {
   // replay
-  await homePage.initiateLogin();
-  await expect(page).toHaveURL(/.*home/);
-  await homePage.createPatient();
-  await homePage.startPatientVisit();
-  await homePage.goToLabOrderForm();
+  await openmrs.goToLabOrderForm();
   await page.getByRole('button', { name: 'Add', exact: true }).click();
   await page.locator('#tab select').selectOption('161447AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
-  await homePage.saveLabOrder();
-  await homePage.goToSENAITE();
+  await openmrs.saveLabOrder();
+  await senaite.goToSENAITE();
   await expect(page).toHaveURL(/.*senaite/);
-  await homePage.searchClientInSENAITE();
-  await homePage.createPartition();
+  await senaite.searchClientInSENAITE();
+  await senaite.createPartition();
   await page.locator('div:nth-child(4) div table tbody tr td.contentcell.Result div span input').fill('Test result: Normal');
-  await homePage.publishLabResults();
+  await senaite.publishLabResults();
   const reviewState = await page.locator('table tbody tr.contentrow.state-published.parent td.contentcell.State span span').textContent();
   await expect(reviewState?.includes('Published')).toBeTruthy();
 
   // verify
   await page.goto(`${O3_URL}`);
-  await homePage.viewTestResults();
+  await openmrs.viewTestResults();
   const testName = await page.locator('div:nth-child(2) >div> div.cds--data-table-container td:nth-child(1)').first();
   await expect(testName).toHaveText('Stool microscopy with concentration');
 
@@ -204,6 +183,6 @@ test('Published free text lab results from SENAITE are viewable in the OpenMRS l
 });
 
 test.afterEach(async ({ page }) => {
-  await homePage.deletePatient();
+  await openmrs.deletePatient();
   await page.close();
 });
